// src/app/Models/TrabajadorBD.ts
export class TrabajadorBD{
  // cedula_trab: string;
  // nom_trab: string;
  // apell_trab: string;
  // tel_trab: string;
  // dir_trab: string;
  // correo_trab: string;
  // salario_trab: number;
  // cod_contrato: string;
  // estado: boolean;
  cedulaTrab: string;
  nomTrab:  string;
  apellTrab: string;
  telTrab: string;
  dirTrab: string;
  correoTrab: string;
  salarioTrab:number;
  codContrato:  string;
  estado:boolean;
}